requirements.txt
altgraph==0.17.4
blinker==1.9.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
reportlab==4.4.3
setuptools==80.9.0
python-dotenv==1.0.1
asgiref==3.8.1
uvicorn[standard]==0.32.0
Werkzeug==3.1.3
requests==2.32.3
